style: hook 분리
