+++ src/utils/dialog/RecordingModal/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useRecordingManager } from './useRecordingManager';
+export { default as useVisualizerManager } from './useVisualizerManager';
